[libromdata] AndroidManifestXML: Parse integers properly based on their data type.
Instead of the bad hack that converts 0 to "false", properly convert
the integer to a string based on the data type. This lets us better
distinguish between actual resource IDs and values that are stored
as integers.

We can remove parseResourceID() from most of the field code, except
for glEsVersion, which still needs to be parsed because it's an
unsigned integer as a string. New function parseUnsignedInteger()
handles this case for both decimal and hexadecimal.

TODO: Parse TYPE_STRING and some other stuff. Most of the string
resource IDs are TYPE_REFERENCE, which requires resource lookups.
TYPE_STRING probably means "look at the string table".