[libromdata] AndroidManifestXML, AndroidResourceReader: Add missing endianness conversion macros to fix APK handling on big-endian.
AndroidManifestXML: Most of it was correct due to the use of LEW(),
except for a few things that accessed Android structs instead of
reading offsets directly.

AndroidResourceReader:

- Change package_id from `unsigned int` to `uint32_t`.

- read_u16(), read_u32(): Optimize.

- Add byteswapping macros where needed.

- alocale: The main value is used as-is without byteswapping.
  The order of the two 16-bit words changes depending on byte order,
  but each word is still stored in BE16.

Tested reading an APK in my Debian 8 PPC qemu VM and it works now.
Previously, it either didn't work or crashed, depending on how badly
the byteswapped values were screwing it up.