AndroidResourceReader: New StringPoolAccessor class that replaces processStringPool().
This class takes the string pool pointer and size, and can be used to
retrieve a string from the string table. The retrieved string is copied
because it might be in UTF-16LE.

While this might not help too much for AndroidResourceReader, since it
ends up loading all the strings anyway for resource type handling, it
will be useful for AndroidManifestXML, which actually does use a similar
format as the resource file. This would replace AndroidManifestXML's
compXmlString() and compXmlStringAt() functions.

Sidenote: Removing the vector<string> objects *does* save a good amount
of memory.