[librpbase] RpWebP.cpp: Add Windows support.
Note that we have to explicitly load libsharpyuv first. Otherwise,
libwebp won't load because it depends on libsharpyuv, and our
DLL restrictions prevent the dynamic linker from finding libsharpyuv.

NOTE: Still not bundling libwebp, but I tested it with a quick and
dirty build of libwebp-1.5.0 compiled with MinGW-w64 and it worked.