[libromdata] AndroidAPK: Add some preliminary (but disabled) code for adaptive icons.
This *was* going to simply draw the foreground drawable, or the
background drawable, without compositing. However, the APK I
tested, Magisk-v24.3.apk, turned out to have SVG drawables.

Adding support for SVG requires pulling in an SVG renderer,
and I'm not going to do that right now.

Not going to drop the code, but it'll be disabled for now.

AndroidResourceReader::findIconHighestDensity(): Added a ".xml"
check, though it's disabled for now.