[librpfile] IRpFile: Remove seek_cur().
We could have used seek_cur() instead of adding `whence` to seek(),
but seek_cur() is less efficient because it has to call tell().

The various seek() implementations can use an internal variable if
it's available instead of calling tell().

seek_cur() was mostly used by CdiReader, but it was also used by
Xbox360_XEX and TGA. Updated them to use seek() and verified that
everything still works correctly.