[scripts] setup?.iss: Check for MSVC Runtime.
setup6.iss: Use IsMsiProductInstalled() to check for the MSI upgrade codes
for the i386, amd64, and arm64 versions.

FIXME: Not sure if we need the i386 version on arm64 systems...

setup5.iss: IsMsiProductInstalled() isn't available, so check for the
msvcp140.dll file.

Note that {sys} has System32 redirection, so it's always the 32-bit
System directory. {syswow64} is the 64-bit System32 on 64-bit, and
the 32-bit System32 on 32-bit.

FIXME: MsgBox() doesn't have clickable links...