[scripts] setup?.iss: Register/unregister the Property Description Schema.
For Release builds, we'll handle it here, so we don't have to create
a C:\Windows\PropDesc directory.

For Debug builds, DllRegisterServer will still register it.

Note that PROPSYS.DLL isn't part of Windows XP by default. (It's
added as part of Windows Search.) Since we're only using the .propdesc
for custom columns, and we can use IColumnProvider for that on XP and
2003, we'll simply skip registration on XP and 2003.