[scripts] package.cmd: Call ISCC.exe to build the Inno Setup packages.
Requires both Inno Setup 5 and Inno Setup 6 in order to build the
XP/2003/Vista and 7/8/10/11 packages.

Removed all usage of cpack.

The PDB files are now explicitly specified and packaged manually.
This does mean I'll have to remember to update package.cmd if
a DLL is changed or added. (Maybe add something to fail if the
file is missing? Right now it doesn't fail...)

TODO: Get the rom-properties version number for %ZIP_PREFIX%.
This was previously obtained using the filenames created by
cpack.