[scripts] setup?.iss: Use built-in functions to determine CPU architecture.
This simplifies the code a bit.

On Inno Setup 5, it now shows a "better" file size for the "Main Files"
component, but the size is 3.3 MB on both Win7/64 and WinXP/32.
It should be smaller on 32-bit.

On Inno Setup 6, it's still showing 49 KB.

Need to figure out a better way to add disk space usage.
May need to use ISPP to scan for files?