[gtk3] Combine the column description data from NautilusColumnProvider.c and NautilusInfoProvider.cpp.
It's now stored in NautilusColumnProvider.c.

rp_nautilus_info_provider_process(): Added a sanity check for the
rp_nautilus_column_provider_column_desc_data[] array size. It's only
run the first time the function is called, and only in debug builds.

TODO: A better way to export it so we can use the compile-time checks
again instead of the debug-only runtime checks?