[scripts] setup6.iss: Rework files to use ISPP to enumerate .dll and .exe files.
It's a bit hackish, since it seems like we can't call the same
ISPP #sub with different #define values (it just emits the same
files), but it seems to work. There's five sets of subroutines,
one for each architecture.

TODO: Get the file size and add it to ExtraDiskSpaceRequired to
get the files to be counted as part of the "Main Files" component.
Currently, the "Check" condition is preventing InnoSetup from
doing that. (This *really* should be fixed in InnoSetup itself...)