[scripts] setup5.iss: Use the ISPP version for the per-architecture DLLs and EXEs.
Note that, for both setup5.iss and setup6.iss, this does result in
rpcli.exe and rp-config.exe for all architectures being installed
in each architecture subdirectory, instead of just one version in
the rom-properties base directory.

This does not match the v2.6 distribution, but it's probably better
to include all of them in order to debug architecture-specific
issues more easily.