[win32] rp-config: Add the Windows 10 ROM chip icon.
From imageres.dll, ICO resource 34.

The 256x256 PNG version was replaced with the modified one from scripts/
that gets rid of the screwy background.

Note that the icon is ~53.5 KB, whereas rp-config.exe (32-bit release
build) is 19 KB. That's a significant size increase, but it's needed
in order to have a proper icon...