[win32] Initial attempt to add 16-bit EXE thumbnailing on 64-bit Windows.
Based on Icon16bitFix:
https://github.com/otya128/Icon16bitFix/blob/master/Icon16bitFix/Icon16bitFix.cpp

Uses the Microsoft Detours library:
https://github.com/microsoft/Detours

Unlike Icon16bitFix, instead of running a standalone program to hook into
Windows Explorer, we're just going to use Detours to hook into programs
that already load the rom-properties shell extension. This is simpler and
doesn't require adding a program to startup, and it reduces the chances
of false-positive virus detection.

Currently, both PrivateExtractIconsW() and PrivateExtractIconsA() are
detoured. The A function simply converts the filename to Unicode and
calls the W function. The W function doesn't actually do anything yet,
so all executable icons are showing up as black squares.

TODO:
- Call the original PrivateExtractIcons() functions and only use our own
  version if it fails.
- Actually extract icons in RP_PrivateExtractIconsW().