[libromdata] EXE: New function loadIconResourceData().
This gets the actual resource data for the requested icon index at the
requested size. The highest color-depth icon is selected.

Export the EXE constructor and loadIconResourceData() so it can be used
by rom-properties.dll.

This function is implemented in a new file, EXE_icons.cpp.

[win32] RP_PrivateExtractIconsW_int(): Call EXE::loadIconResourceData(),
then use CreateIconFromResourceEx() to create the icon.

Windows 3.1 calc.exe is now thumbnailed on Windows 10 amd64.

TODO:
- rom-properties.dll might not be loaded by explorer initially, since
  we're not registering as an icon or thumbnail handler for EXEs.
  Viewing the EXE properties and then refreshing gets it to load.
- Handle requests for multiple icons.
- Handle Windows 1.x/2.x executables.