[zstd] Restructure so we can include contrib/seekable_format/.
The seekable format library is needed for Z3DS compressed ROMs.

Added a CMakeLists.txt to build seekable_format as a static library.

[extlib] CMakeLists.txt: Build zstd seekable_format.

Rework SET_EXTLIB_PROPERTIES() such that it excludes targets if
SKIP_INSTALL_ALL is set. This was needed to get it to skip seekable_Format
right now, since nothing's using it, but it might break things...