[libromdata] Improve .zip file handling for AndroidAPK and J2ME.
RomDataFactory will now open the .zip file using MiniZip and search
for the required files for each class. If found, it will pass the
unzFile to the subclass directly.

AndroidAPK and J2ME now have an additional constructor that takes an
unzFile argument for an opened .zip file. (If nullptr, the .zip file
will be opened by the subclass.)

The 1-argument constructor now forwards to the 2-argument constructor
with a nullptr unzFile.

This should improve performance in some cases by only having to
initialize MiniZip once.