[win32] Move .propdesc registration back to DllRegisterServer.
InnoSetup only registers it in WOW6432Node, since it runs as a
32-bit (i386) executable. In order to register it for 64-bit,
PSRegisterPropertySchema() needs to run in the 64-bit DLL.

RP_PropertyStore_Private::GetPropertyDescriptionSchemaFilename():
New function that checks for the .propdesc file in the DLL's current
or parent directories. Previously, this retrieved "C:\\Windows\\PropDesc"
and was only used in debug builds.

DllRegisterServer(): Don't unregister the .propdesc before registering
it. Otherwise, the 64-bit DLL will unregister the 32-bit schema, and
vice-versa.

TODO: Copy the .propdesc file to the binary directory for testing
purposes.