[libromdata/tests] RomDataViewTest: Fix a memory leak in the GTK2 and GTK3 versions.
gtk_widget_destroy() doesn't work properly if the widget is floating.
Use g_object_ref_sink() to sink the floating reference, then
g_clear_object() to unreference and destroy it.

This basically uses the GTK4 method for GTK2 and GTK3, and it seems
to work correctly.

NOTE: RomDataViewTest_gtk3 still fails if built with AddressSanitizer
due to a false-positive memory leak in fontconfig.